Rename source to sources
