+++ investd/__main__.py
@@ -3,7 +3,7 @@ from pathlib import Path
 
 import click
 
-from .source import ingest_sources
+from .sources import ingest_sources
 
 log = logging.getLogger("investd")
 
@@ -15,13 +15,13 @@ def investd():
 
 @click.command(name="ingest-sources")
 @click.option("--output", default="data/transactions/tx.csv", help="Output path")
-def ingest_sources_cli(output: Path):
+def ingest_sources_cmd(output: Path):
     df_tx = ingest_sources()
     log.info(f"Writing {output}")
     df_tx.to_csv(output, index=False)
 
 
-investd.add_command(ingest_sources_cli)
+investd.add_command(ingest_sources_cmd)
 
 
 if __name__ == "__main__":
